examples/src/backend/requirements.txt
Flask==2.3.0
Flask-CORS==4.0.0
mysql-connector-python==8.0.33
python-dotenv==1.0.0
bcrypt==4.0.1
PyJWT==2.8.0
requests==2.31.0
python-dateutil==2.8.2
click==8.1.3
Werkzeug==2.3.0
Jinja2==3.1.2
MarkupSafe==2.1.1
itsdangerous==2.1.2
cryptography==40.0.1
